refactor(#1696): split config.rs + lib.rs to keep the file-size ratchet green
The fix rounds grew four over-threshold files past main's line counts. Three are
fixed by splitting, per the campsite rule rather than an override flag:

* `cqlite-core/src/config.rs` 978 -> 722: `Config::validate` and the rules it
  enforces move to `config_validate.rs`, the JSON entry points to
  `config_json.rs` (epic #1116). Both are cohesive responsibilities whose
  commentary was dominating the type definitions.
* `cqlite-core/src/lib.rs` 817 -> 760: the crate-root inline `mod tests` moves to
  `lib_tests.rs` (#1135).
* `cqlite-cli/src/config.rs` back to main's count: the new seam's docstring is
  trimmed to a pointer at the `removed_keys` module doc, which already carries
  the ordering rationale in full.
* `reader/mod.rs` net zero: the fraction is validated AT the resolver call site
  instead of in a preceding block, with the reasoning in the resolver's own doc
  and in `validated_direct_io_memory_fraction`. This drops the (nice but not
  required) "before any file I/O" placement; the enforcement point is unchanged
  and the end-to-end test still proves the reader rejects rather than clamps.

Also: dropped the outer `///` doc on `pub mod config_removed_keys;` — an outer
doc on a `pub mod` merges with the file's `//!` block and resolves intra-doc
links in the PARENT scope, so four links pointing at that module's own items
were silently broken (`no item named Removed in module cqlite_core`). Same for
the CLI's declaration. Verified with `cargo doc`.

Shared warning wording is now surface-neutral ("the configuration still loads"),
since one formatter serves both the file and the JSON surfaces; the assertions
that pinned the old file-specific phrasing were updated with it.
